refactor: data:lectureevaluation:viewer, reporter 통합
