+++ app-compose/build.gradle.kts
@@ -24,8 +24,7 @@ dependencies {
   implementation(projects.data.openmajor)
   implementation(projects.data.timetable)
   implementation(projects.data.lectureevaluation.editor)
-  implementation(projects.data.lectureevaluation.viewer)
-  implementation(projects.data.lectureevaluation.report)
+  implementation(projects.data.lectureevaluation.viewerreporter)
   implementation(projects.data.lectureevaluation.my)
   implementation(projects.data.user)
   implementation(projects.data.notice)
