--- data/lectureevaluation/report/src/main/java/com/suwiki/data/lectureevaluation/report/datasource/RemoteExamReportDataSource.kt
@@ -1,9 +1,0 @@
-package com.suwiki.data.lectureevaluation.report.datasource
-
-interface RemoteExamReportDataSource {
-
-  suspend fun reportExam(
-    evaluateIdx: Long,
-    content: String = "",
-  )
-}
