--- data/lectureevaluation/report/src/main/java/com/suwiki/data/lectureevaluation/report/datasource/RemoteLectureReportDataSource.kt
@@ -1,9 +1,0 @@
-package com.suwiki.data.lectureevaluation.report.datasource
-
-interface RemoteLectureReportDataSource {
-
-  suspend fun reportLecture(
-    evaluateIdx: Long,
-    content: String = "",
-  )
-}
