--- data/lectureevaluation/report/src/test/java/com/suwiki/data/lectureevaluation/report/ExampleUnitTest.kt
@@ -1,16 +1,0 @@
-package com.suwiki.data.lectureevaluation.report
-
-import junit.framework.TestCase.assertEquals
-import org.junit.Test
-
-/**
- * Example local unit test, which will execute on the development machine (host).
- *
- * See [testing documentation](http://d.android.com/tools/testing).
- */
-class ExampleUnitTest {
-  @Test
-  fun addition_isCorrect() {
-    assertEquals(4, 2 + 2)
-  }
-}
