--- data/lectureevaluation/viewer/build.gradle.kts
@@ -1,8 +1,0 @@
-@Suppress("DSL_SCOPE_VIOLATION") // TODO: Remove once KTIJ-19369 is fixed
-plugins {
-  id("suwiki.android.data")
-}
-
-android {
-  namespace = "com.suwiki.data.lectureevalution.viewer"
-}
