+++ data/lectureevaluation/viewerreporter/proguard-rules.pro
