+++ data/lectureevaluation/viewerreporter/src/main/AndroidManifest.xml
