+++ data/lectureevaluation/viewerreporter/src/main/java/com/suwiki/data/lectureevaluation/viewerreporter/datasource/RemoteExamProviderDataSource.kt
@@ -1,0 +1,12 @@
+package com.suwiki.data.lectureevaluation.viewerreporter.datasource
+
+import com.suwiki.core.model.lectureevaluation.LectureDetailExamData
+
+interface RemoteExamProviderDataSource {
+  suspend fun buyExam(lectureId: Long)
+
+  suspend fun getLectureDetailExam(
+    lectureId: Long,
+    page: Int,
+  ): LectureDetailExamData
+}
