+++ data/lectureevaluation/viewerreporter/src/main/java/com/suwiki/data/lectureevaluation/viewerreporter/datasource/RemoteExamReportDataSource.kt
@@ -1,0 +1,9 @@
+package com.suwiki.data.lectureevaluation.viewerreporter.datasource
+
+interface RemoteExamReportDataSource {
+
+  suspend fun reportExam(
+    evaluateIdx: Long,
+    content: String = "",
+  )
+}
