+++ data/lectureevaluation/viewerreporter/src/main/java/com/suwiki/data/lectureevaluation/viewerreporter/datasource/RemoteLectureReportDataSource.kt
@@ -1,0 +1,9 @@
+package com.suwiki.data.lectureevaluation.viewerreporter.datasource
+
+interface RemoteLectureReportDataSource {
+
+  suspend fun reportLecture(
+    evaluateIdx: Long,
+    content: String = "",
+  )
+}
