+++ data/lectureevaluation/viewerreporter/src/test/java/com/suwiki/data/lectureevaluation/viewerreporter/ExampleUnitTest.kt
@@ -1,0 +1,16 @@
+package com.suwiki.data.lectureevaluation.viewerreporter
+
+import junit.framework.TestCase.assertEquals
+import org.junit.Test
+
+/**
+ * Example local unit test, which will execute on the development machine (host).
+ *
+ * See [testing documentation](http://d.android.com/tools/testing).
+ */
+class ExampleUnitTest {
+  @Test
+  fun addition_isCorrect() {
+    assertEquals(4, 2 + 2)
+  }
+}
