+++ remote/lectureevaluation/viewerreporter/src/main/java/com/suwiki/remote/lectureevaluation/viewerreporter/datasource/RemoteExamReportDataSourceImpl.kt
@@ -1,6 +1,6 @@
 package com.suwiki.remote.lectureevaluation.viewerreporter.datasource
 
-import com.suwiki.data.lectureevaluation.report.datasource.RemoteExamReportDataSource
+import com.suwiki.data.lectureevaluation.viewerreporter.datasource.RemoteExamReportDataSource
 import com.suwiki.remote.lectureevaluation.viewerreporter.api.ExamReportApi
 import com.suwiki.remote.lectureevaluation.viewerreporter.request.ReportExamRequest
 import javax.inject.Inject
