+++ remote/lectureevaluation/viewerreporter/src/main/java/com/suwiki/remote/lectureevaluation/viewerreporter/datasource/RemoteLectureReportDataSourceImpl.kt
@@ -1,6 +1,6 @@
 package com.suwiki.remote.lectureevaluation.viewerreporter.datasource
 
-import com.suwiki.data.lectureevaluation.report.datasource.RemoteLectureReportDataSource
+import com.suwiki.data.lectureevaluation.viewerreporter.datasource.RemoteLectureReportDataSource
 import com.suwiki.remote.lectureevaluation.viewerreporter.api.LectureReportApi
 import com.suwiki.remote.lectureevaluation.viewerreporter.request.ReportLectureRequest
 import javax.inject.Inject
