+++ settings.gradle.kts
@@ -42,8 +42,7 @@ include(":local:timetable")
 include(":local:user")
 include(":data:openmajor")
 include(":data:timetable")
-include(":data:lectureevaluation:report")
-include(":data:lectureevaluation:viewer")
+include(":data:lectureevaluation:viewerreporter")
 include(":data:lectureevaluation:editor")
 include(":data:lectureevaluation:my")
 include(":data:user")
